+++ iqshell/sandbox/injection_rule/operations/update.go
@@ -37,7 +37,7 @@ func Update(info UpdateInfo) {
 		return
 	}
 
-	client, err := sbClient.NewSandboxClient()
+	client, err := sbClient.NewInjectionRuleClient()
 	if err != nil {
 		sbClient.PrintError("%v", err)
 		return